version 1.6, - new search, infinite map handling, etc

+++ Dockerfile
@@ -2,7 +2,8 @@
 FROM node:lts-alpine
 # create directory, install xml editing tool and setup non-root user
 RUN adduser -h /app/server -s /bin/sh -D -u 1001 server && \
-    apk add xmlstarlet su-exec
+    apk add xmlstarlet su-exec && \
+    npm i -g npm@10.9.1
 USER server
 WORKDIR /app/server
 # copy useful trakman files and entrypoint command
@@ -13,6 +14,7 @@ COPY --chown=1001:1001 ./Plugins.ts ./package.json ./tsconfig.json ./CHANGELOG.m
 COPY --chown=1001:1001 --chmod=0755 ./docker_run.sh ./docker_run.sh
 # download dedicated server and remove unnecessary files
 RUN wget -O serv.zip http://files2.trackmaniaforever.com/TrackmaniaServer_2011-02-21.zip && \
+    if [ "$(sha256sum serv.zip |  cut -d ' ' -f 1)" != "dd021f49c3d58d45ee09f333015bdb35b19243a38fa65f202ca8a88fb6550c0b" ]; then exit 1; fi && \
     unzip serv.zip && \
     rm -r serv.zip CommandLine.html ListCallbacks.html ListMethods.html \
     Readme_Dedicated.html RemoteControlExamples TrackmaniaServer.exe manialink_dedicatedserver.txt
@@ -30,8 +32,9 @@ RUN mv GameData/Config/dedicated_cfg.txt dedicated_cfg.txt.bk && \
 # expose volumes
 VOLUME /app/server/GameData/Config
 VOLUME /app/server/GameData/Tracks
+VOLUME /app/server/Logs
 VOLUME /app/server/trakman
 VOLUME /app/server/.pm2/logs
 # set user back to root to be able to chown volumes in the run script
 USER root
-CMD /app/server/docker_run.sh
+CMD ["/app/server/docker_run.sh"]
